+++ src/messaging/messaging-api-request-internal.ts
@@ -18,8 +18,8 @@ import { FirebaseApp } from '../firebase-app';
 import {
   HttpMethod, AuthorizedHttpClient, HttpRequestConfig, HttpError, HttpResponse,
 } from '../utils/api-request';
-import { createFirebaseError, getErrorCode } from './messaging-errors';
-import { SubRequest, BatchRequestClient } from './batch-request';
+import { createFirebaseError, getErrorCode } from './messaging-errors-internal';
+import { SubRequest, BatchRequestClient } from './batch-request-internal';
 import { SendResponse, BatchResponse } from './messaging-types';
 import { getSdkVersion } from '../utils/index';
